Fixed validateAadat? I think
